add description of steamstoreapi to understand what include
