+++ src/crowlexecuter/__init__.py
@@ -1,0 +1,4 @@
+__author__ = "TOSUKUi"
+
+from . import steamfirst
+
