+++ src/main.py
@@ -1,0 +1,5 @@
+import execute
+
+if __name__ == "__main__":
+
+    execute.executefirst.execute()
